+++ Sources/CohesionKit/IdentityMap.swift
@@ -3,56 +3,108 @@ import Combine
 
 
 /// Framework main class.
-/// Store and access publishers referencing `Identifiable` objects to have realtime updates on them.
+/// Store and access publishers referencing objects to have realtime updates on them.
 /// Memory is automatically released when objects have no observers
 public class IdentityMap {
-    var map:[String:[String:Any]] = [:]
+    
+    /// in-memory storages. Storages are deallocated automatically (unless aliased)
+    var values: [String:Any] = [:]
+    /// access stored values by names. `aliases` have their own storage and values are forwarded to it. Therefore
+    /// when aliased a value won't be released automatically
+    /// example: "current_user" to find the current user
+    /// - Parameter storage: the alias storage
+    /// - Parameter token: token on alias `storage.publisher` the storage alive
+    var aliases: [String:(storage: Any, token: AnyCancellable?)] = [:]
 
-    /// Create an identity map with a compare function determining when data should be considered as stale and replaced.
-    /// - Parameter isStale: this function is used when calling `update` to determine whether or not received data should
-    /// replace the existing one. First parameter is existing data, second one is new one
     public init() {
     
     }
     
-    func remove<Model>(for model: Model.Type, id: Any) {
-        self[model, id: id] = nil
+    func remove<Model>(for type: Model.Type, id: Any) {
+        self[type, id: id] = nil
+    }
+    
+    /// Remove an alias from the identity map.
+    /// You should call this method to release any data strongly referenced by `alias`
+    public func remove(alias: String) {
+        aliases[alias] = nil
     }
 
-    /// Access the storage for Model given its type and id
+    /// Access and set the storage for Model given its type and id
     subscript<Model>(type: Model.Type, id id: Any) -> Storage<Model>? {
-        get { map["\(type)"]?[String(describing: id)] as? Storage<Model> }
-        set { map["\(type)", default: [:]][String(describing: id)] = newValue }
+        get { values[key(for: type, id: id)] as? Storage<Model> }
+        set { values[key(for: type, id: id)] = newValue }
+    }
+    
+    /// Access the storage and initialize it if value is not present
+    subscript<Model>(type: Model.Type, id id: Any, init default: @autoclosure () -> Storage<Model>) -> Storage<Model> {
+        guard let storage = values[key(for: type, id: id)] as? Storage<Model> else {
+            let storage = `default`()
+            
+            values[key(for: type, id: id)] = storage
+            
+            return storage
+        }
+        
+        return storage
+    }
+    
+    /// Register a storage under a given alias. Values will be accessible using this alias
+    func register<Model>(alias: String?, storage valueStorage: Storage<Model>) {
+        guard let key = alias else {
+            return
+        }
+        
+        let storage = self.storage(aliased: key) as Storage<Model>
+        
+        storage.merge(valueStorage.publisher, modifiedAt: Date().stamp)
+        
+        aliases[key] = (storage: storage, token: storage.publisher.sink(receiveValue: { _ in }))
+    }
+    
+    /// Find an aliased storage. Be careful: This is storage from `values` but from `aliases`
+    func storage<Model>(aliased key: String) -> Storage<Model> {
+        if let storage = aliases[key]?.storage as? Storage<Model> {
+            return storage
+        }
+        
+        let alias = (storage: Storage<Model> { }, token: AnyCancellable?.none)
+        
+        aliases[key] = alias
+        
+        return alias.storage
+    }
+    
+    private func key<Model>(for type: Model.Type, id: Any) -> String {
+        "\(type)-\(id)"
     }
     
     /// Add or update an element in the storage with its new value.
     ///
     /// You usually use this method in conjunction with `publisherIfPresent(for:id:)`
-    /// - Returns: a Publisher emitting new values for the element. Object stay in memory as long as someone is using the publisher, otherwise it is realeased from the identity map
+    /// - Returns: a Publisher emitting new values for the element. Object stay in memory as long as someone is using the publisher (or if aliased), otherwise it is realeased from the identity map
     /// - Parameter element: the element to add or update
     /// - Parameter relation: Describe the element and how it will be inserted into the identity map.
+    /// - Parameter alias: a string key which can be used to find back the element without having its id
     /// - Parameter modifiedAt: If value is higher than previous update then the element will be updated. Otherwise changes will be ignored.
     public func store<Element, ID: Hashable>(
         _ element: Element,
         using relation: Relation<Element, ID>,
+        alias: String? = nil,
         modifiedAt: Stamp = Date().stamp
     ) -> AnyPublisher<Element, Never> {
         let id = element[keyPath: relation.idKeyPath]
         
-        guard let publisher = storeIfPresent(element, using: relation, modifiedAt: modifiedAt) else {
-            let storage = Storage<Element>(id: id, identityMap: self)
+        let storage = self[Element.self, id: id, init: Storage<Element>(id: id, identityMap: self)]
+        
+        storage.merge(
+            recursiveStore(element, using: relation, modifiedAt: modifiedAt),
+            modifiedAt: modifiedAt
+        )
+        
+        register(alias: alias, storage: storage)
 
-            self[Element.self, id: id] = storage
-
-            storage.merge(
-                recursiveStore(element, using: relation, modifiedAt: modifiedAt),
-                modifiedAt: modifiedAt
-            )
-
-            return storage.publisher
-        }
-
-        return publisher
+        return storage.publisher
     }
     
     /// Add or update multiple elements at once into the storage
@@ -69,26 +121,23 @@ public class IdentityMap {
     
     /// Update element in the storage only if it's already in it. Otherwise discard the changes.
     ///
-    /// You usually use this method in conjunction with `publisher(for:id:)` which will always create a storage for the
+    /// You usually use this method in conjunction with `publisher(using:id:)` which will always create a storage for the
     /// element with specified id.
     /// - SeeAlso:
-    /// `IdentityMap.store(_,relation:,modifiedAt:)`
+    /// `IdentityMap.store(_:relation:alias:modifiedAt:)`
     @discardableResult
     public func storeIfPresent<Element, ID: Hashable>(
         _ element: Element,
         using relation: Relation<Element, ID>,
+        alias: String? = nil,
         modifiedAt: Stamp = Date().stamp
     ) -> AnyPublisher<Element, Never>? {
-        guard let storage = self[Element.self, id: element[keyPath: relation.idKeyPath]] else {
+        
+        guard self[Element.self, id: element[keyPath: relation.idKeyPath]] != nil else {
             return nil
         }
-
-        storage.merge(
-            recursiveStore(element, using: relation, modifiedAt: modifiedAt),
-            modifiedAt: modifiedAt
-        )
-
-        return storage.publisher
+        
+        return store(element, using: relation, alias: alias, modifiedAt: modifiedAt)
     }
     
     /// Return a publisher emitting event when receiving update for `id`.
@@ -100,23 +149,24 @@ public class IdentityMap {
         using relation: Relation<Element, ID>,
         id: ID
     ) -> AnyPublisher<Element, Never> {
-        guard let storage = self[Element.self, id: id] else {
-            let storage = Storage<Element>(id: id, identityMap: self)
-
-            self[Element.self, id: id] = storage
-
-            return storage.publisher
-        }
-
+        let storage = self[Element.self, id: id, init: Storage<Element>(id: id, identityMap: self)]
+        
         return storage.publisher
+    }
+    
+    /// Return a publisher emitting event when receiving update on `alias`
+    public func publisher<Element>(for element: Element.Type, aliased alias: String) -> AnyPublisher<Element, Never> {
+        storage(aliased: alias).publisher
     }
 
     /// Return element with matching `id` if an object with such `id` was previously inserted
-    public func get<Element, ID: Hashable>(
-        using relation: Relation<Element, ID>,
-        id: ID
-    ) -> Element? {
+    public func get<Element, ID: Hashable>(using relation: Relation<Element, ID>, id: ID) -> Element? {
         self[Element.self, id: id]?.value
+    }
+    
+    /// Return element matching `alias`
+    public func get<Element>(for element: Element.Type, aliased alias: String) -> Element? {
+        storage(aliased: alias).value
     }
     
     private func recursiveStore<Element, ID: Hashable>(