+++ Tests/CohesionKitTests/IdentityMapTests.swift
@@ -13,7 +13,7 @@ class IdentityMapTests: XCTestCase {
         XCTAssertNotNil(map.get(for: Entity.self, id: Entity.hello.id))
     }
 
-    func test_getForId_entityWasAdded_allPublishersWereUnsubscribed_entityIsNotStored() {
+    func test_getForId_valueIsAdded_valueIsCancelled_itReturnNil() {
         let map = IdentityMap()
         var cancellables: Set<AnyCancellable> = []
 
@@ -46,7 +46,45 @@ class IdentityMapTests: XCTestCase {
 
         XCTAssertEqual(receivedValue, Entity.hello)
     }
+    
+    func test_getAliased_valueWasStoreWithAlias_itRetunValue() {
+        let map = IdentityMap()
+        var cancellables: Set<AnyCancellable> = []
+        
+        map.store(Entity.hello, alias: "test_alias")
+            .sink(receiveValue: { _ in })
+            .store(in: &cancellables)
+        
+        XCTAssertNotNil(map.get(for: Entity.self, aliased: "test_alias"))
+    }
+    
+    func test_getAliased_valueIsCancelled_itReturnValue() {
+        let map = IdentityMap()
+        var cancellables: Set<AnyCancellable> = []
+        
+        map.store(Entity.hello, alias: "test_alias")
+            .sink(receiveValue: { _ in })
+            .store(in: &cancellables)
+        
+        cancellables.removeAll()
+        
+        XCTAssertNotNil(map.get(for: Entity.self, aliased: "test_alias"))
+    }
 
+    func test_getAliased_valueIsCancelled_aliasIsRemoved_itReturnNil() {
+        let map = IdentityMap()
+        var cancellables: Set<AnyCancellable> = []
+        
+        map.store(Entity.hello, alias: "test_alias")
+            .sink(receiveValue: { _ in })
+            .store(in: &cancellables)
+        
+        cancellables.removeAll()
+        
+        map.remove(alias: "test_alias")
+        
+        XCTAssertNil(map.get(for: Entity.self, aliased: "test_alias"))
+    }
 }
 
 private struct Entity: Identifiable, Equatable {